Fixed error, humidity list for userId now contains decimals instead of integers

+++ src/main/java/org/terraeyes/terraeyesdb/dataAccess/humidity/HumidityDaoImpl.java
@@ -39,7 +39,7 @@ public class HumidityDaoImpl extends DaoConnection implements HumidityDao
             resultSet.getInt("id"),
             resultSet.getString("eui"),
             resultSet.getString("timestamp"),
-            resultSet.getInt("humidity")
+            resultSet.getBigDecimal("humidity")
         );
 
         measurements.add(measurement);
